Geri kalan tüm Dto ların validator işlemleri bitirildi

+++ src/AnkaYazilim.OnMuhasebe.Application.Contracts/DTOs/Hizmetler/CreateHizmetDto.cs
@@ -4,7 +4,7 @@ public class CreateHizmetDto:IEntityDto
 {
     public string Kod { get; set; }
     public string Ad { get; set; }
-    public byte? KdvOrani { get; set; }
+    public int? KdvOrani { get; set; }
     public decimal? BirimFiyat { get; set; }
     public string Barkod { get; set; }
     public Guid? BirimId { get; set; }
