+++ src/AnkaYazilim.OnMuhasebe.Application.Contracts/DTOs/Stoklar/CreateStokDto.cs
@@ -4,7 +4,7 @@ public class CreateStokDto:IEntityDto
 {
     public string Kod { get; set; }
     public string Ad { get; set; }
-    public byte? KdvOran { get; set; }
+    public int? KdvOran { get; set; }
     public string Barkod { get; set; }
     public decimal? SatisFiyat { get; set; }
     public decimal? AlisFiyat { get; set; }
